upgrade to msvc 19.28.29336
- mostly just fix a whole ton of warnings due to implicit f32
  conversions
- stb_ds: add a missing function signature (not sure how this was
  compiling before this???...)

+++ code-utilities/kgtAudioMixer.cpp
@@ -140,7 +140,8 @@ internal void
 				const u8 rawSoundChannel = static_cast<u8>(
 					c % tRawSound.channelCount);
 				const SoundSample sample = static_cast<SoundSample>(
-					*(rawSoundSampleBlockStart + rawSoundChannel) * 
+					static_cast<f32>(
+						*(rawSoundSampleBlockStart + rawSoundChannel)) * 
 					tracks[t].volumeRatio);
 				*(sampleBlockStart + c) += sample;
 			}
